선착순 이벤트 참여 관련 메소드 리팩토링 (#4)
* refactor: 선착순 이벤트 참여 DB 접근을 위한 BlockingQueue가 10000개까지 변수를 수용할 수 있도록 수정

* docs: 주석 추가

* refactor: 락을 걸고 OrderApplyCount를 가져오는 메소드를 새로운 Interface에서 수행하도록 수정

* refactor: CurrentOrderEvent 테스트 수정

CurrentOrderEvent 단위 테스트가 객체 자신의 OrderApplyCount에 의존하도록 수정

* refactor: CurrentOrderEventManagerService -> OrderEventFromServerMemoryService로 수정

적합한 이름으로 변경

* refactor: OrderApplyCount가 저장 되어있는 곳(서버, DB)에 따른 Service가 분리되도록 수정

* test: OrderEventFromServerMemoryServiceTest의 단위 테스트의 OrderApplyCountService에 대한 의존성 제거
