+++ order/src/main/java/com/watermelon/server/order/service/OrderEventCommandService.java
@@ -11,7 +11,7 @@ import com.watermelon.server.order.domain.OrderEvent;
 import com.watermelon.server.order.repository.OrderEventRepository;
 
 
-import com.watermelon.server.orderResult.service.CurrentOrderEventManageService;
+import com.watermelon.server.orderResult.service.OrderEventFromServerMemoryService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,17 +23,17 @@ public class OrderEventCommandService {
 
     private final OrderEventRepository orderEventRepository;
     private final OrderEventWinnerService orderEventWinnerService;
-    private final CurrentOrderEventManageService currentOrderEventManageService;
+    private final OrderEventFromServerMemoryService orderEventFromServerMemoryService;
 
 
 
     public OrderEventCommandService(
             OrderEventRepository orderEventRepository,
             OrderEventWinnerService orderEventWinnerService,
-            CurrentOrderEventManageService currentOrderEventManageService) {
+            OrderEventFromServerMemoryService orderEventFromServerMemoryService) {
         this.orderEventRepository = orderEventRepository;
         this.orderEventWinnerService = orderEventWinnerService;
-        this.currentOrderEventManageService = currentOrderEventManageService;
+        this.orderEventFromServerMemoryService = orderEventFromServerMemoryService;
         findOrderEventToMakeInProgress();
     }
 
@@ -47,13 +47,13 @@ public class OrderEventCommandService {
     @Transactional
     public Long findOrderEventToMakeInProgress(){
         List<OrderEvent> orderEvents = orderEventRepository.findAll();
-        if(orderEvents.isEmpty()) return currentOrderEventManageService.getCurrentOrderEventId();// 이벤트 없을시 스킵
+        if(orderEvents.isEmpty()) return orderEventFromServerMemoryService.getCurrentOrderEventId();// 이벤트 없을시 스킵
         for(OrderEvent orderEvent : orderEvents){
             //현재 이벤트중 시간에 맞는 이벤트가 있다면 현재 이벤트로 설정한다
             if(orderEvent.isTimeInEventTime(LocalDateTime.now())){
-                this.currentOrderEventManageService.refreshOrderEventInProgress(orderEvent);
+                this.orderEventFromServerMemoryService.refreshOrderEventInProgress(orderEvent);
             }
         }
-        return currentOrderEventManageService.getCurrentOrderEventId();
+        return orderEventFromServerMemoryService.getCurrentOrderEventId();
     }
 }
