+++ order/src/main/java/com/watermelon/server/orderResult/service/OrderApplyCountService.java
@@ -1,0 +1,9 @@
+package com.watermelon.server.orderResult.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface OrderApplyCountService {
+
+    public boolean isOrderApplyCountAddable(long orderApplyCountId,int eachMaxWinnerCount);
+}
